Migrate distributed read osquery endpoint
